Move lint deps to dev and add FlatCompat
