+++ eslint.config.js
@@ -1,0 +1,4 @@
+const { FlatCompat } = require('@eslint/eslintrc');
+const compat = new FlatCompat({ baseDirectory: __dirname });
+
+module.exports = compat.config(require('./.eslintrc.js'));
