+++ package.json
@@ -30,16 +30,15 @@
     "babel-plugin-import-rename": "^1.0.1",
     "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
     "@babel/plugin-transform-react-jsx": "^7.22.0",
-    "@babel/preset-env": "^7.22.0"
-  },
-  "optionalDependencies": {
+    "@babel/preset-env": "^7.22.0",
     "ajv": "^8.12.0",
     "@babel/eslint-parser": "^7.22.0",
     "eslint": "^8.44.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-react": "^7.32.0",
-    "prettier": "^3.0.0"
+    "prettier": "^3.0.0",
+    "@eslint/eslintrc": "^2.1.3"
   },
   "scripts": {
     "build-css": "sass --load-path=./src --load-path=./node_modules src/:src/",
