+++ src/assets/jss/material-kit-react/components/footerStyle.jsx
@@ -1,4 +1,4 @@
-import { container, primaryColor } from "../material-kit-react.jsx";
+import { container, primaryColor } from "../../material-kit-react.jsx";
 
 const footerStyle = {
   block: {
