+++ src/assets/jss/material-kit-react/views/analyzerPage.jsx
@@ -1,4 +1,4 @@
-import { container, title } from "../material-kit-react.jsx";
+import { container, title } from "../../material-kit-react.jsx";
 
 const landingPageStyle = {
   container: {
