+++ src/assets/jss/material-kit-react/views/landingPageSections/productStyle.jsx
@@ -1,4 +1,4 @@
-import { title } from "../../material-kit-react.jsx";
+import { title } from "../../../material-kit-react.jsx";
 
 const productStyle = {
   section: {
