+++ src/assets/jss/material-kit-react/views/landingPageSections/workStyle.jsx
@@ -1,4 +1,4 @@
-import { title } from "../../material-kit-react.jsx";
+import { title } from "../../../material-kit-react.jsx";
 
 const workStyle = {
   section: {
